05/30/2022 Mon

+++ utils/utils.py
@@ -9,35 +9,40 @@ import time
 import itertools
 import os
 from difflib import SequenceMatcher
+import xlrd
+from datetime import timedelta
 
 class Global(object):
     root_path = 'C:/Users/qcells/Desktop/ATP/US_ATP_BI'
     driver= '{ODBC Driver 17 for SQL Server}'  
-    server = 'qcells-us-atp-db-server.database.windows.net'
-    database = 'us-qcells-atp-db'
-    username = 'qcells'
-    password = '{Asdqwe123!@#}'   
+    atp_server = 'qcells-us-atp-db-server.database.windows.net'
+    atp_database = 'us-qcells-atp-db'
+    atp_username = 'qcells'
+    atp_password = '{Asdqwe123!@#}'   
+    qsp_server = '11.0.40.144'
+    qsp_database = 'master'
+    qsp_username = 'qspdev_user'
+    qsp_password = '{QsxdRqsp123!#*}'   
 
 class DB_Utils():
     def connect_qspdb(self):
         try:
-            self.conn = pyodbc.connect('DRIVER='+Global.driver
-                                    +';SERVER=tcp:'+Global.server
-                                    +';PORT=1433;DATABASE='+Global.database
-                                    +';UID='+Global.username
-                                    +';Pwd='+Global.password
-                                    +';Encrypt=yes;TrustServerCertificate=no;Connection Timeout=30;')
-            self.cursor = self.conn.cursor()
+            self.qsp_conn = pyodbc.connect('DRIVER='+Global.driver
+                                    +';SERVER=tcp:'+Global.qsp_server
+                                    +';PORT=1433;DATABASE='+Global.qsp_database
+                                    +';UID='+Global.qsp_username
+                                    +';Pwd='+Global.qsp_password)
+            self.qsp_cursor = self.conn.cursor()
         except:
             raise ConnectionError('ConnectionError:AZURE DB CONNECTION FAILED')
 
     def connect_azuredb(self):
         try:
             self.conn = pyodbc.connect('DRIVER='+Global.driver
-                                    +';SERVER=tcp:'+Global.server
-                                    +';PORT=1433;DATABASE='+Global.database
-                                    +';UID='+Global.username
-                                    +';Pwd='+Global.password
+                                    +';SERVER=tcp:'+Global.atp_server
+                                    +';PORT=1433;DATABASE='+Global.atp_database
+                                    +';UID='+Global.atp_username
+                                    +';Pwd='+Global.atp_password
                                     +';Encrypt=yes;TrustServerCertificate=no;Connection Timeout=30;')
             self.cursor = self.conn.cursor()
         except:
@@ -51,7 +56,7 @@ class DB_Utils():
             col_names = [item[0] for item in self.cursor.description]
             return pd.DataFrame(row, columns=col_names)
         except:
-            raise ConnectionError('ConnectionError:RPA FAILED TO INSERT DATA TO AZURE DB TABLES')
+            raise ConnectionError('ConnectionError:RPA FAILED TO FETCH DATA TO AZURE DB TABLES')
 
     def insert_dataframe(self, tablename, df, writer = False):
         self.connect_azuredb()
@@ -90,7 +95,6 @@ class QSP_MasterCode(DB_Utils):
         sql_atp = '''
             select count(1) from atp_master_code
             '''
-        
         qsp_master_code_qty = self.fetch_data(sql_qsp).values()[0][0]
         atp_master_code_qty = self.fetch_data(sql_qsp).values()[0][0]
         if qsp_master_code_qty - atp_master_code_qty > 0:
@@ -103,8 +107,8 @@ class QSP_MasterCode(DB_Utils):
             sql = 'select * from qsp_master_code'
             master_data = self.fetch_data(self)
             self.insert_dataframe(master_data)
-
         self.conn.close()
+        self.isUpadted = False
 
 class Master_Reset(DB_Utils):
     def check_isadmin(self):
@@ -119,26 +123,43 @@ class Master_Reset(DB_Utils):
         try:
             self.master_df = pd.ExcelFile(Global.root_path + '/data/master.xlsx',engine='openpyxl')
             for sheet in self.master_df.sheet_names:
-                if sheet != 'US_CITY_CODE':
+                if (sheet != 'US_CITY_CODE') and (sheet != 'CW_CALENDER') and (sheet != 'ITEM_CODE_MASTER') and (sheet != 'POWER_CLASS_DIST'):
                     sql = 'DELETE FROM {}'.format(sheet)
                     self.cursor.execute(sql)
-            self.conn.commit()
-        except:
-            raise ValueError('MasterResetError:WHILE RESETTING MASTER TABLES, THERE WAS AN ERROR.')
+                    self.conn.commit()
+        except Exception as e :
+            raise ValueError('MasterResetError:WHILE RESETTING MASTER TABLES, THERE WAS AN ERROR. \n', str(e))
 
     def push_data(self):
         try:
             print('[EVENT] NEW MASTER DATA IS BEING SAVED !')
             for sheet_name in self.master_df.sheet_names:
-                if sheet_name != 'US_CITY_CODE':
+                if (sheet_name != 'US_CITY_CODE') and (sheet_name != 'CW_CALENDER') and (sheet_name != 'ITEM_CODE_MASTER') and (sheet_name != 'POWER_CLASS_DIST'):
                     df = pd.ExcelFile(self.file_path,engine='openpyxl').parse(sheet_name)
-                    df = df.loc[:, ~df.columns.str.contains('^Unnamed')]
+                    # df = df.loc[:, ~df.columns.str.contains('^Unnamed')]
+                    df = df.dropna(how='all', axis='columns')
                     df.dropna(inplace = True, how = 'all')
                     df.fillna('', inplace = True)
+                    if sheet_name == 'POWER_CLASS_DIST':
+                        df = self.unpivot_powerclassdist(df)
+                    if sheet_name == 'CW_CALENDER':
+                        df['Friday'] = df['Friday'].astype('str')
+                    if sheet_name == 'TARIFF_INFO':
+                        df = self.unpivot_taiffinfo(df)
                     self.insert_dataframe(sheet_name, df)
             print('\n')
-        except:
-            raise ValueError('MasterResetError:WHILE PUSHING DATA TO MASTER TABLES, THERE WAS AN ERROR.')
+        except Exception as e:
+            raise ValueError('MasterResetError:WHILE PUSHING DATA TO MASTER TABLES, THERE WAS AN ERROR. \n', str(e))
+
+    def unpivot_powerclassdist(self, df):
+        power_df = pd.melt(df, id_vars=['FactoryCd','Product_Name', 'SAP_UNIQUE_CODE', 'Power_Class', 'Segment'], value_vars=df.columns[5:])
+        power_df = power_df.rename(columns = {'variable' : 'YYYYMM', 'value' : 'Distribution'})
+        power_df.YYYYMM = power_df.YYYYMM.astype('str')
+        return power_df
+        
+    def unpivot_taiffinfo(self, df):
+        df = df.melt(id_vars = ['Year', 'Month', 'YYYYMM'], value_vars=df.columns[3:]).rename(columns = {'variable' : 'FactoryCd', 'value' : 'TariffCd'})
+        return df
 
     def master_reset_main(self, sender_addr, file_path = Global.root_path+'/data/master.xlsx'): 
         self.file_path = file_path
@@ -155,26 +176,38 @@ class ETL_Utils(DB_Utils):
         self.target_sheet_list = list(set(self.sheet_info['sheetlist'].tolist()))
         self.conn.close()
 
+    def order_status_column_filter(self):
+        self.df.columns = [col.split('(')[0].rstrip() for col in self.df.columns]
+        self.df.rename(columns = {'Task Name' : 'CustomerPO_Num'
+                            , 'Task ID' : 'CR_Num'}, inplace = True)
+
     def read_rpa_excel_file(self, path, engine, sheet_name):
         time.sleep(1)
         skiprows = 0
+        if self.excel_name == 'ORDER_STATUS_REPORT':
+            skiprows = 2
         while True:
             try:
                 print('[WARNING] {} / {} EXCEL FILE SKIP ROWS: {}'.format(path, sheet_name, skiprows))
-                df = pd.read_excel(path
+                self.df = pd.read_excel(path
                                     , engine=engine
                                     , sheet_name = sheet_name
                                     , skiprows= skiprows)
-                df = df.loc[:, ~df.columns.str.contains('^Unnamed')]
+                self.df = self.df.loc[:, ~self.df.columns.str.contains('^Unnamed')]
                 break
             except Exception as e:
                 skiprows = skiprows + 1
+                if skiprows > 5:
+                    raise ValueError('PLEASE CHECK OUT THE COLUMN LINE IN {} EXCEL FILE'.format(path)) 
                 print(e)
                 pass
-        df = df.dropna(how = 'all')
-        df = df[self.sheet_info.targetcolumns.tolist()]
-        df.columns = [col.strip() for col in df.columns]
-        return df
+        self.df = self.df.dropna(how = 'all')
+        self.df.columns = [col.strip() for col in self.df.columns]
+
+        if self.excel_name == 'ORDER_STATUS_REPORT':
+            self.order_status_column_filter()
+        self.df = self.df[self.sheet_info.targetcolumns.tolist()]
+        return self.df
 
     def read_excel(self, path = Global.root_path):
         try:
@@ -242,10 +275,10 @@ class ETL_Utils(DB_Utils):
         else:
             x = x[:8]
         try:
-            return datetime.strftime(datetime.strptime(x, '%Y%m%d'), '%Y-%m-%d')
+            return datetime.strftime(datetime.strptime(x, '%m%d%Y'), '%Y-%m-%d')
         except:
             try:
-                return datetime.strftime(datetime.strptime(x, '%m%d%Y'), '%Y-%m-%d')
+                return datetime.strftime(datetime.strptime(x, '%Y%m%d'), '%Y-%m-%d')
             except:
                 return ''
                 pass
@@ -302,6 +335,31 @@ class ETL_Utils(DB_Utils):
             else:
                 pass
 
+    def read_date(self, date):
+        return xlrd.xldate.xldate_as_datetime(date, 0)
+
+    def data_prod_plan_clean(self):
+        self.excel_name = 'GA_PROD_PLAN'
+        self.df = pd.read_excel(self.saved_path
+                            , engine='pyxlsb'
+                            , sheet_name = 'Revised_Detail PSI (item code)'
+                            , skiprows= 4)
+        self.df = self.df.dropna(how='all', axis='columns')
+        self.df = self.df[self.df['Curr/Futu Prod.'] !='X']
+        self.df = self.df[self.df['MODEL'].notna()]
+        self.df = self.df[self.df['category'] == 'W/H IN (생산)']
+        self.df = self.df.drop(['category', 'Grade','Cell type', 'J.Box', 'Curr/Futu Prod.'], axis = 1)
+        self.df = self.df.reset_index(drop = True)
+        self.df = self.df.melt(id_vars = ['MODEL', 'Power', 'Item code'])
+        self.df['variable']= pd.to_datetime(self.df['variable'].apply(self.read_date), errors='coerce')
+        self.df = self.df.dropna()
+        self.df['variable'] = self.df['variable'].apply(lambda x:self.yyyymmdd_datetime(str(x)))
+        self.df = self.df.rename(columns = {'MODEL' : 'ProductName',
+                            'Power' : 'Power_Class',
+                            'Item code' : 'Item_Code',
+                            'variable' : 'Product_Plan_Date', 
+                            'value' : 'MW'})
+    
     def clean_data_text(self, func):
         try:
             target_columns_text = self.sheet_info[self.sheet_info['datatype'] == 'TEXT']['targetcolumns'].tolist()
@@ -339,12 +397,13 @@ class ETL_Utils(DB_Utils):
         self.insert_dataframe(self.excel_name, self.df, writer = True)
 
 class ETL_Pipelines(ETL_Utils, DB_Utils):
-    def __init__(self, saved_path, mail_category, check_date = datetime.now().strftime('%Y-%m-%d'), writer = 'SYSTEM'):
+    def __init__(self, saved_path='', mail_category='', check_date = datetime.now().strftime('%Y-%m-%d'), writer = 'SYSTEM', manual = False):
         self.saved_path = saved_path
         self.mail_category = mail_category
         self.check_date_sql = check_date # => dataframe의 updated_date를 참조
         self.writer = writer
-        # self.main()
+        if manual == False:
+            self.main()
 
     def ETL_Basic_Process(self):
         self.read_excel(path = self.saved_path)
@@ -377,6 +436,12 @@ class ETL_Pipelines(ETL_Utils, DB_Utils):
         self.meta_sheet_info(excel_name = 'INBOUND_DOMESTIC')
         self.ETL_Basic_Process()
 
+    def GA_PROD_PLAN_ETL(self):
+        # self.read_excel(path = self.saved_path)        
+        self.data_prod_plan_clean()
+        self.Check_file_is_updated()
+        self.insert_dataframe('GA_PROD_PLAN', self.df, writer = True)
+
     def main(self):
         print('[EVENT] DATA ETL PIPELINE JUST BEGAN : {} !'.format(self.mail_category))
         if self.mail_category == 'INBOUND_OVERSEA':
@@ -389,6 +454,8 @@ class ETL_Pipelines(ETL_Utils, DB_Utils):
             self.ON_HAND_ETL()
         if self.mail_category == 'ORDER_STATUS_REPORT':
             self.ORDER_STATUS_REPORT_ETL()
+        if self.mail_category == 'GA_PROD_PLAN':
+            self.GA_PROD_PLAN_ETL()
         self.ETL_error_check()
 
 class Email_Utils(Master_Reset, ETL_Utils):
@@ -445,25 +512,30 @@ class Email_Utils(Master_Reset, ETL_Utils):
         self.atts = list(itertools.chain(*atts))                
         self.sender_addr = self.sender_mailaddr_extract(self.i)
 
-    def category_replace(self):
-        if self.mail_category_parse.lower() == 'inbound_oversea':
-            self.mail_category_parse = 'INBOUND_OVERSEA'
-        elif self.mail_category_parse.lower() == 'inbound_domestic':
-            self.mail_category_parse = 'INBOUND_DOMESTIC'
-        elif self.mail_category_parse.lower() == 'outbound':
-            self.mail_category_parse = 'OUT_BOUND'
-        elif self.mail_category_parse.lower() == 'onhand':
-            self.mail_category_parse = 'ON_HAND'
-        elif self.mail_category_parse.lower() == 'osr':
-            self.mail_category_parse = 'ORDER_STATUS_REPORT'
+    def category_replace(self, mail_category_parse):
+        #MAIL LIST SHEET에서 참조함.
+        if mail_category_parse.lower() == 'inbound_oversea':
+            mail_category_parse = 'INBOUND_OVERSEA'
+        elif mail_category_parse.lower() == 'inbound_domestic':
+            mail_category_parse = 'INBOUND_DOMESTIC'
+        elif mail_category_parse.lower() == 'outbound':
+            mail_category_parse = 'OUT_BOUND'
+        elif mail_category_parse.lower() == 'onhand':
+            mail_category_parse = 'ON_HAND'
+        elif mail_category_parse.lower() == 'osr':
+            mail_category_parse = 'ORDER_STATUS_REPORT'
+        elif mail_category_parse.lower() == 'ga_prod_plan':
+            mail_category_parse = 'GA_PROD_PLAN'
         else:
-            raise ValueError('ValueError:THERE IS NO CATEGORY IN THE LIST, ex) inbound -> IN_BOUND')
+            mail_category_parse = 'None'
+        return mail_category_parse
+            # raise ValueError('ValueError:THERE IS NO CATEGORY IN THE LIST, ex) inbound -> IN_BOUND')
 
     def extract_request_update_mail_info(self):
         try:
             title_parsed = self.i.subject.lower().split('/')
-            self.mail_category_parse = re.sub(r'[^a-zA-Z]', '', title_parsed[1])
-            self.category_replace()
+            mail_category_parse = re.sub(r'[^a-zA-Z]', '', title_parsed[1])
+            self.mail_category_parse = self.category_replace(mail_category_parse)
             self.Req_date = datetime.now().strftime('%Y-%m-%d')
             if len(title_parsed) >= 3:
                 if len(title_parsed[2]) == 8:
@@ -480,7 +552,7 @@ class Email_Utils(Master_Reset, ETL_Utils):
         similarity = 0
         similarity_attach = 0
         isdomainsame = False
-        self.mail_category_parse = 'None'
+        mail_category_parse = 'None'
         for idx, whitelist in enumerate(whitelist_dict):
             if similarity <= SequenceMatcher(None, whitelist['Mailtitle'], mail_title).ratio():
                 similarity = SequenceMatcher(None, whitelist['Mailtitle'], mail_title).ratio()
@@ -491,10 +563,9 @@ class Email_Utils(Master_Reset, ETL_Utils):
                     if similarity_attach <= SequenceMatcher(None, whitelist['ExcelName'], attch_name).ratio():
                         similarity_attach = SequenceMatcher(None, whitelist['ExcelName'], attch_name).ratio()
                         max_ssim_excel = whitelist['ExcelName']
-                        self.mail_category_parse = whitelist['MailCategory']
-                        self.category_replace()
-                        
-                if (mail_domain == max_ssim_domian) & (similarity > self.similarity_threshold):
+                        mail_category_parse = whitelist['MailCategory']
+                        self.mail_category_parse= self.category_replace(mail_category_parse)                
+                if (mail_domain == max_ssim_domian) & (similarity > self.similarity_threshold) & (similarity_attach > self.similarity_threshold):
                     isdomainsame = True
                     break
                 else:
